requirements.txt
lxml
tqdm
pyyaml
pymysql
requests
sqlalchemy
sqlacodegen
DrissionPage
fake_useragent
